% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_diffprep.R
\name{me_diffprep}
\alias{me_diffprep}
\title{Preparing for Metabolic Analysis}
\usage{
me_diffprep(
  data = data,
  info = info,
  group1 = "R",
  group2 = "NR",
  filter = FALSE
)
}
\arguments{
\item{data}{An expression matrix with genes as row names and samples as column names.}

\item{info}{A data frame containing sample group information.
The first column must be named \code{id}, and it should exactly match the column names (i.e., sample names) of the expression matrix.
The second column must be named \code{type}, representing binary classification labels.}

\item{group1}{Group1}

\item{group2}{Group2}

\item{filter}{FALSE}
}
\value{
A data frame prepared for downstream metabolic analysis.
}
\description{
Preparing for Metabolic Analysis
}
\examples{
me_diffprep(data = data, info = info, group1 = "R", group2 = "NR", filter = FALSE)
}
