% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_leading_edge.R
\name{me_leading_edge}
\alias{me_leading_edge}
\title{Title Leading edge analysis}
\usage{
me_leading_edge(input = diff, geneset = list)
}
\arguments{
\item{input}{The me_diffprep results.}

\item{geneset}{The metabolism-related gene set that you interested in.}
}
\value{
The leading edge results.
}
\description{
Title Leading edge analysis
}
\examples{
me_leading_edge(input = diff, geneset = list)
}
