% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_leading_edge_plot.R
\name{me_leading_edge_plot}
\alias{me_leading_edge_plot}
\title{Title Leading edge analysis visualization.}
\usage{
me_leading_edge_plot(
  input = diff,
  geneset = list,
  leading = TRUE,
  select.gene = FALSE,
  select.name = select
)
}
\arguments{
\item{input}{The me_diffprep results.}

\item{geneset}{The metabolism-related gene set that you interested in.}

\item{leading}{TRUE}

\item{select.gene}{TRUE / FALSE}

\item{select.name}{the metabolism-related genes that you interested in.}
}
\value{
A leading edge plot.
}
\description{
Title Leading edge analysis visualization.
}
\examples{
me_leading_edge_plot(input = diff, geneset = list, leading = TRUE, select.gene = FALSE, select.name = select)
}
