% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_lollipop_plot.R
\name{me_lollipop_plot}
\alias{me_lollipop_plot}
\title{Title}
\usage{
me_lollipop_plot(input = meta)
}
\arguments{
\item{input}{The me_pathwayanalysis result.}
}
\value{
A lollipop plot.
}
\description{
Title
}
\examples{
me_lollipop_plot(input = meta)
}
