% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_volcano_plot.R
\name{me_volcano_plot}
\alias{me_volcano_plot}
\title{Title MetabolismExplorer visualization: Volcano plot}
\usage{
me_volcano_plot(
  input = meta,
  output = name,
  thres.p = 0.05,
  thres.nes = 1,
  topn = 3,
  marker = select,
  label.size = 5,
  width = 8,
  height = 7
)
}
\arguments{
\item{input}{The result of me_pathwayanalysis.}

\item{output}{The name of the output file}

\item{thres.p}{threshold for p value}

\item{thres.nes}{thershold for NES}

\item{topn}{highlight the top n altered metabolic signatures}

\item{marker}{the metabolic signature you interested in}

\item{label.size}{label size}

\item{width}{figure width}

\item{height}{figure hight}
}
\value{
A volcano plot showing the enrichment results.
}
\description{
Title MetabolismExplorer visualization: Volcano plot
}
\examples{
me_volcano_plot(input = meta, output = name, thres.p = 0.05, thres.nes = 1, topn = 3, marker = select, label.size = 5, width = 8, height = 7)
}
